fix: address review comments (#36)
