+++ src/widgets/home-page/index.ts
@@ -1,0 +1,1 @@
+export { HomePage } from "./ui/home-page";
